Fill table before calling openPopup() and position arrow automatically
MDN suggests using the popupshowing event to change the contents of
popups before they're shown, but unfortunately arrow panels use the same
event to position the arrow. The end result is that the location of the
arrow is picked before the box is sized to the new content, so the arrow
ends up in the wrong place. Fix by filling the contents out before even
calling openPopup().

Automatic arrow positioning is needed for the toolbar button, to make sure
the panel doesn't appear in silly places; this code tries to place the
popup in the direction of the middle of the browser window.